filin_mate/requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0 
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
pycodestyle==2.9.1
PyJWT==2.4.0
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.4
